base64 encode google credentials for CircleCI
CircleCI does not handle multiline environment variables well, so
base64 encoding the google application credentials.
(https://circleci.com/docs/google-auth/ has an example of base64
encoding the google credential json file and storing it as an
environment variable).
Then we can decode this environment variable and use it in the Ruby
google auth library as per:
https://github.com/google/google-auth-library-ruby/issues/78