Add default Ruby sinatra devcontainer settings
We want to be able to run the application in a [Visual Studio Code
Docker devcontainer][1] (including on [GitHub Codespaces][2]), rather
than having to install all the software locally.

This commit adds the [default Ruby sinatra devcontainer settings][3].

The application will not run on the devcontainer yet, as we will need
to customise the container settings to install the additional software
we need (e.g. Redis).  We will apply these customisations in subsequent
commits.

[1]: https://code.visualstudio.com/docs/remote/containers
[2]: https://github.com/features/codespaces
[3]: https://github.com/microsoft/vscode-dev-containers/tree/master/containers/ruby-sinatra