Add YAML linting check on all pushes (#100)
* Add YAML lint check on git pushes

The check uses yamllint[1], and runs via the Reviewdog[2]
GitHub Action[3]

[1] https://yamllint.readthedocs.io/
[2] https://github.com/reviewdog/reviewdog
[3] https://github.com/reviewdog/reviewdog#github-actions

* Allow yaml lines of any length

The default was to report an error for any line over 80 characters,
which is too restrictive.  We now have it set to warn rather than error,
and only if the line is over 120 characters rather than 80.

* Fix yaml linting errors

These linting errors we fix here should all be cosmetic only, as the
application was building fine before we started linting.