Use language server to improve Ruby IDE experience
The VS Code [Ruby extension highly recommends using the language server]
[1], which is why we have switched to it in this commit.

It integrates with RuboCop, so we have also removed the separate RuboCop
extension.

We are using the default language server settings suggested by the
extension's homepage (except we use a globally installed RuboCop gem,
rather than bundling it with the project).

[1]: https://marketplace.visualstudio.com/items?itemName=rebornix.Ruby