Enable remote containers and Codespaces dev (#102)
* Add default Ruby sinatra devcontainer settings

We want to be able to run the application in a [Visual Studio Code
Docker devcontainer][1] (including on [GitHub Codespaces][2]), rather
than having to install all the software locally.

This commit adds the [default Ruby sinatra devcontainer settings][3].

The application will not run on the devcontainer yet, as we will need
to customise the container settings to install the additional software
we need (e.g. Redis).  We will apply these customisations in subsequent
commits.

[1]: https://code.visualstudio.com/docs/remote/containers
[2]: https://github.com/features/codespaces
[3]: https://github.com/microsoft/vscode-dev-containers/tree/master/containers/ruby-sinatra

* Set devcontainer Ruby version to match Gemfile's

* Ensure gpg is available to git to sign commits

* Set Visual Studio Code as the git editor

* Suppress Dockerfile unpinned gem versions warning

We will pin the gem versions in a later commit.  Not doing it in this
commit as want to take the smallest steps possible to get the
application running successfully in a Docker devcontainer.

* Bump Ruby to 2.5.8 so it can run in devcontainer

The version before this commit (2.5.3) is not available as a Docker
devcontainer image.

* Install RuboCop on Docker devcontainer

[RuboCop][1] is the predominant Ruby linter. We have already been using
it on this project for a long time, so it definitely needs to be on the
devcontainer.

[1]: https://github.com/rubocop-hq/rubocop

* Add ruby-rubocop devcontainer extension

[Ruby RuboCop extension homepage][1]

Also ensured that the RuboCop execution path and settings location are
set correctly via the `devcontainer.json` file.

[1]: https://github.com/misogi/vscode-ruby-rubocop

* Update location of RuboCop line length check

The new location was introduced in [RuboCop 0.78.0][1], via [this PR][2]

[1]: https://github.com/rubocop-hq/rubocop/blob/master/CHANGELOG.md#0780-2019-12-18
[2]: https://github.com/rubocop-hq/rubocop/pull/7542

* Install application gems upon container creation

* Install Redis on devcontainer

Redis is needed for the application to run locally.

We set Redis to start every time the container starts or restarts.

* Install Heroku to run app locally on devcontainer

To run the app locally on the devcontainer we use [Heroku Local][1].
We install Heroku Local using the [standalone method][2].

[1]: https://devcenter.heroku.com/articles/heroku-local
[2]: https://devcenter.heroku.com/articles/heroku-cli#standalone-installation

* Make the web app available outside devcontainer

We use [Heroku Local][1] to run the web app, so we forward Heroku
Local's default port of 5000 so we can access it outside the container.

[1]: https://devcenter.heroku.com/articles/heroku-local

* Stop installing Sinatra on devcontainer

There is no need, as it is declared in the Gemfile as one of the
project's gems.

* Stop installing Nodejs on the devcontainer

We don't use Node in this project.

* Use language server to improve Ruby IDE experience

The VS Code [Ruby extension highly recommends using the language server]
[1], which is why we have switched to it in this commit.

It integrates with RuboCop, so we have also removed the separate RuboCop
extension.

We are using the default language server settings suggested by the
extension's homepage (except we use a globally installed RuboCop gem,
rather than bundling it with the project).

[1]: https://marketplace.visualstudio.com/items?itemName=rebornix.Ruby

* Document the new local development workflow

Added documentation for remote containers and Codespaces.