Install Redis on separate Docker container
We now use Docker Compose to install Redis as a separate Docker service
on its own container.  This is a proper use of Docker, allowing our
data service to be separate to our appliation.  It fixes an issue where
the Redis process was not running in the background and was preventing
the postcreatecommand from finishing properly.

Our compose file is based on [this example repo][1]
(although our app uses neither Python or Postgres, the principle of an
app linked to a service is the same)

[1]: https://github.com/microsoft/vscode-dev-containers/tree/master/containers/python-3-postgres

Also see
https://code.visualstudio.com/docs/remote/create-dev-container#_using-docker-compose