Notify us of non-automatic version updates
This is a hack, to get Dependabot to notify us about version updates for
dependencies that it cannot update automatically.

We never want this hack GitHub Action to be triggered (because the
Dependabot doesn't need the action to be triggered in order to carry out
it's version updating), so we have set it to only ever trigger on a
branch name which we will never have.

The PR that Dependabot creates will only have the update for this
`dependabot_hack` file, so we'll have to then manually update the
version wherever it is specified too.  We should do this in the same
PR that Dependabot has created.

https://docs.github.com/en/github/administering-a-repository/keeping-your-actions-up-to-date-with-github-dependabot