Fix Redis gem version problem
Prior to this commit we were using the verbose pessimistic constraint
version specifier for the Redis gem (documented [here][1], along with
the shorthand version).  Dependabot was not picking
up the pessimistic constraint and was [wrongly creating a PR to bump us
to the Redis 4.x gem][2], which we don't want to happen.

Fixed by using the [`~>` pessimistic constraint shortcut][3].

[1]: https://guides.rubygems.org/patterns/#pessimistic-version-constraint
[2]: https://github.com/johnboyes/hotcustard-payments/pull/112
[3]: https://thoughtbot.com/blog/rubys-pessimistic-operator