Remove Ubuntu update scheduler (#261)
It turns out that GitHub Actions only supports Ubuntu LTS versions,
which are every 2 years.  GitHub Actions cron scheduler functionality
does not support setting a workflow to be every 2 years, so this commit
removes the scheduled creation of a GitHub Issue for this.  Instead the
repository owner has their own personal (non-GitHub) reminder to update
the Ubuntu version every 2 years on 15 July.